divided main into multuple pieces to improve readability and maintainability.
