+++ src/main.rs
@@ -3,10 +3,20 @@ use clap::Parser;
 use std::{env, fs, path::Path, sync::Arc};
 use tokio::sync::Semaphore;
 use tokio_postgres::NoTls;
+use crate::types::{IndexInfo, SharedTableTracker};
+use crate::checks::perform_reindexing_checks;
+use crate::connection::{create_connection_with_session_parameters, set_session_parameters};
+use crate::index_operations::{get_indexes_in_schema, reindex_index_with_client};
+
 mod logging;
 mod queries;
 mod save;
 mod schema;
+mod types;
+mod checks;
+mod deadlock;
+mod connection;
+mod index_operations;
 
 #[derive(Parser, Debug)]
 #[command(author, version, about = "PostgreSQL Index Reindexer - Reindexes all indexes in a specific schema or table", long_about = None)]
@@ -98,740 +108,6 @@ struct Args {
     /// Log file path (default: reindexer.log in current directory)
     #[arg(short = 'l', long, default_value = "reindexer.log")]
     log_file: String,
-}
-
-#[derive(Debug)]
-struct IndexInfo {
-    schema_name: String,
-    index_name: String,
-    index_type: String,
-}
-
-#[derive(Debug, Clone)]
-struct ReindexingCheckResults {
-    active_vacuum: bool,
-    active_pgreindexer: bool,
-    inactive_replication_slots: bool,
-    sync_replication_connection: bool,
-}
-
-#[derive(Debug)]
-struct SharedTableTracker {
-    tables_being_reindexed: std::collections::HashMap<String, String>, // table_name -> index_name
-}
-
-// check if there is an already running pgreindexer process
-async fn get_running_pgreindexer(client: &tokio_postgres::Client) -> Result<bool> {
-    let rows = client
-        .query(queries::GET_RUNNING_PGREINDEXER, &[])
-        .await
-        .context("Failed to query running pgreindexer")?;
-    Ok(rows.len() > 0)
-}
-
-// check the active vacuums
-async fn get_active_vacuum(client: &tokio_postgres::Client) -> Result<bool> {
-    let rows = client
-        .query(queries::GET_ACTIVE_VACUUM, &[])
-        .await
-        .context("Failed to query active vacuums")?;
-    Ok(rows.len() > 0)
-}
-
-// check the inactive replication slots
-async fn get_inactive_replication_slots(client: &tokio_postgres::Client) -> Result<bool> {
-    let rows = client
-        .query(queries::GET_INACTIVE_REPLICATION_SLOT_COUNT, &[])
-        .await
-        .context("Failed to query inactive replication slots")?;
-    let inactive_replication_slot_count: i64 = rows.first().unwrap().get(0);
-    Ok(inactive_replication_slot_count > 0)
-}
-
-// check the sync replication connection
-async fn get_sync_replication_connection(client: &tokio_postgres::Client) -> Result<bool> {
-    let rows = client
-        .query(queries::GET_SYNC_REPLICATION_CONNECTION_COUNT, &[])
-        .await
-        .context("Failed to query sync replication connection")?;
-    let sync_replication_connection_count: i64 = rows.first().unwrap().get(0);
-    Ok(sync_replication_connection_count > 0)
-}
-
-async fn check_and_handle_deadlock_risk(
-    client: &tokio_postgres::Client,
-    schema_name: &str,
-    index_name: &str,
-    shared_tracker: &Arc<tokio::sync::Mutex<SharedTableTracker>>,
-    logger: &logging::Logger,
-) -> Result<()> {
-    // Get the table name for this index
-    let table_name = match get_table_name_for_index(client, schema_name, index_name).await {
-        Ok(name) => name,
-        Err(e) => {
-            logger.log(
-                logging::LogLevel::Warning,
-                &format!("[DEBUG] Failed to get table name for index {}.{}: {}. Proceeding without deadlock check.", 
-                    schema_name, index_name, e),
-            );
-            return Ok(());
-        }
-    };
-    let full_table_name = format!("{}.{}", schema_name, table_name);
-
-    logger.log(
-        logging::LogLevel::Info,
-        &format!(
-            "[DEBUG] Checking deadlock risk for index {}.{} on table {}",
-            schema_name, index_name, full_table_name
-        ),
-    );
-
-    let mut retry_count = 0;
-    const MAX_RETRIES: u32 = 12; // 1 hour max (12 * 5 minutes)
-
-    loop {
-        // Check shared tracker for currently reindexed tables
-        let current_tables = {
-            let tracker = shared_tracker.lock().await;
-            tracker
-                .tables_being_reindexed
-                .keys()
-                .cloned()
-                .collect::<Vec<String>>()
-        };
-
-        logger.log(
-            logging::LogLevel::Info,
-            &format!(
-                "[DEBUG] Current tables being reindexed: {:?}",
-                current_tables
-            ),
-        );
-
-        // Check if our table is already being reindexed
-        if current_tables.contains(&full_table_name) {
-            retry_count += 1;
-            if retry_count > MAX_RETRIES {
-                logger.log(
-                    logging::LogLevel::Error,
-                    &format!(
-                        "[DEBUG] Maximum retries exceeded for {}.{}. Proceeding anyway.",
-                        schema_name, index_name
-                    ),
-                );
-                break;
-            }
-
-            logger.log(
-                logging::LogLevel::Warning,
-                &format!("[DEBUG] Potential deadlock detected! Table {} is already being reindexed. Waiting 5 minutes... (retry {}/{})", 
-                    full_table_name, retry_count, MAX_RETRIES),
-            );
-
-            // Wait 5 minutes
-            tokio::time::sleep(tokio::time::Duration::from_secs(300)).await;
-
-            logger.log(
-                logging::LogLevel::Info,
-                &format!(
-                    "[DEBUG] Retrying reindex for {}.{} after 5 minute wait",
-                    schema_name, index_name
-                ),
-            );
-
-            // Continue the loop to check the tracker again after waiting
-            continue;
-        }
-
-        // No deadlock risk, add our table to the tracker and break out of the loop
-        {
-            let mut tracker = shared_tracker.lock().await;
-            tracker
-                .tables_being_reindexed
-                .insert(full_table_name.clone(), index_name.to_string());
-            logger.log(
-                logging::LogLevel::Info,
-                &format!("[DEBUG] Added table {} to reindex tracker", full_table_name),
-            );
-        }
-        break;
-    }
-
-    logger.log(
-        logging::LogLevel::Info,
-        &format!(
-            "[DEBUG] No deadlock risk detected for {}.{}",
-            schema_name, index_name
-        ),
-    );
-
-    Ok(())
-}
-
-async fn remove_table_from_tracker(
-    client: &tokio_postgres::Client,
-    schema_name: &str,
-    index_name: &str,
-    shared_tracker: &Arc<tokio::sync::Mutex<SharedTableTracker>>,
-    logger: &logging::Logger,
-) -> Result<()> {
-    // Get the table name for this index
-    let table_name = match get_table_name_for_index(client, schema_name, index_name).await {
-        Ok(name) => name,
-        Err(e) => {
-            logger.log(
-                logging::LogLevel::Warning,
-                &format!("[DEBUG] Failed to get table name for index {}.{}: {}. Cannot remove from tracker.", 
-                    schema_name, index_name, e),
-            );
-            return Ok(());
-        }
-    };
-    let full_table_name = format!("{}.{}", schema_name, table_name);
-
-    // Remove from shared tracker
-    {
-        let mut tracker = shared_tracker.lock().await;
-        tracker.tables_being_reindexed.remove(&full_table_name);
-        logger.log(
-            logging::LogLevel::Info,
-            &format!(
-                "[DEBUG] Removed table {} from reindex tracker",
-                full_table_name
-            ),
-        );
-    }
-
-    Ok(())
-}
-
-// Perform all reindexing checks once and return results
-async fn perform_reindexing_checks(
-    client: &tokio_postgres::Client,
-) -> Result<ReindexingCheckResults> {
-    let active_vacuum = get_active_vacuum(client).await?;
-    let active_pgreindexer = get_running_pgreindexer(client).await?;
-    let inactive_replication_slots = get_inactive_replication_slots(client).await?;
-    let sync_replication_connection = get_sync_replication_connection(client).await?;
-
-    Ok(ReindexingCheckResults {
-        active_vacuum,
-        active_pgreindexer,
-        inactive_replication_slots,
-        sync_replication_connection,
-    })
-}
-
-async fn get_indexes_in_schema(
-    client: &tokio_postgres::Client,
-    schema_name: &str,
-    table_name: Option<&str>,
-    max_size_gb: u64,
-) -> Result<Vec<IndexInfo>> {
-    let query = if let Some(_table) = table_name {
-        queries::GET_INDEXES_IN_SCHEMA_WITH_TABLE
-    } else {
-        queries::GET_INDEXES_IN_SCHEMA
-    };
-
-    let rows = if let Some(table) = table_name {
-        client
-            .query(query, &[&schema_name, &table, &(max_size_gb as i64)])
-            .await
-    } else {
-        client
-            .query(query, &[&schema_name, &(max_size_gb as i64)])
-            .await
-    }
-    .context("Failed to query indexes in schema")?;
-
-    let mut indexes = Vec::new();
-
-    for row in rows {
-        let index = IndexInfo {
-            schema_name: row.get(0),
-            index_name: row.get(2),
-            index_type: row.get(4),
-        };
-        indexes.push(index);
-    }
-
-    Ok(indexes)
-}
-
-async fn get_index_size(
-    client: &tokio_postgres::Client,
-    schema_name: &str,
-    index_name: &str,
-) -> Result<i64> {
-    let rows = client
-        .query(queries::GET_INDEX_SIZE, &[&schema_name, &index_name])
-        .await
-        .context("Failed to query index size")?;
-
-    if let Some(row) = rows.first() {
-        let size: i64 = row.get(0);
-        Ok(size)
-    } else {
-        Err(anyhow::anyhow!(
-            "Index {}.{} not found",
-            schema_name,
-            index_name
-        ))
-    }
-}
-
-async fn get_table_name_for_index(
-    client: &tokio_postgres::Client,
-    schema_name: &str,
-    index_name: &str,
-) -> Result<String> {
-    let query = r#"
-        SELECT tablename 
-        FROM pg_indexes 
-        WHERE schemaname = $1 AND indexname = $2
-    "#;
-
-    let rows = client
-        .query(query, &[&schema_name, &index_name])
-        .await
-        .context("Failed to query table name for index")?;
-
-    if let Some(row) = rows.first() {
-        let table_name: String = row.get(0);
-        Ok(table_name)
-    } else {
-        Err(anyhow::anyhow!(
-            "Table name not found for index {}.{}",
-            schema_name,
-            index_name
-        ))
-    }
-}
-
-async fn validate_index_integrity(
-    client: &tokio_postgres::Client,
-    schema_name: &str,
-    index_name: &str,
-) -> Result<bool> {
-    let rows = client
-        .query(
-            queries::VALIDATE_INDEX_INTEGRITY,
-            &[&schema_name, &index_name],
-        )
-        .await
-        .context("Failed to query index integrity")?;
-
-    if let Some(row) = rows.first() {
-        let is_valid: bool = row.get(2);
-        let is_ready: bool = row.get(3);
-        let is_live: bool = row.get(4);
-
-        // Index is considered healthy if it's valid, ready, and live
-        Ok(is_valid && is_ready && is_live)
-    } else {
-        // If no rows returned, the index doesn't exist or there's an issue
-        Ok(false)
-    }
-}
-
-async fn reindex_index_with_client(
-    client: Arc<tokio_postgres::Client>,
-    schema_name: String,
-    index_name: String,
-    index_type: String,
-    index_num: usize,
-    total_indexes: usize,
-    verbose: bool,
-    skip_inactive_replication_slots: bool,
-    skip_sync_replication_connection: bool,
-    reindexing_results: Arc<ReindexingCheckResults>,
-    shared_tracker: Arc<tokio::sync::Mutex<SharedTableTracker>>,
-    logger: Arc<logging::Logger>,
-) -> Result<()> {
-    logger.log_index_start(
-        index_num,
-        total_indexes,
-        &schema_name,
-        &index_name,
-        &index_type,
-    );
-
-    logger.log(
-        logging::LogLevel::Info,
-        &format!(
-            "[DEBUG] Starting reindex process for {}.{}",
-            schema_name, index_name
-        ),
-    );
-
-    // Get before size
-    logger.log(
-        logging::LogLevel::Info,
-        &format!(
-            "[DEBUG] Getting before size for {}.{}",
-            schema_name, index_name
-        ),
-    );
-    let before_size = get_index_size(&client, &schema_name, &index_name).await?;
-    logger.log(
-        logging::LogLevel::Info,
-        &format!(
-            "[DEBUG] Before size for {}.{}: {} bytes",
-            schema_name, index_name, before_size
-        ),
-    );
-
-    let reindex_sql = format!(
-        "REINDEX INDEX CONCURRENTLY \"{}\".\"{}\"",
-        schema_name, index_name
-    );
-    logger.log(
-        logging::LogLevel::Info,
-        &format!("[DEBUG] SQL to execute: {}", reindex_sql),
-    );
-
-    // check if the index is invalid before reindexing
-    logger.log(
-        logging::LogLevel::Info,
-        &format!(
-            "[DEBUG] Validating index integrity for {}.{}",
-            schema_name, index_name
-        ),
-    );
-    let index_is_valid = validate_index_integrity(&client, &schema_name, &index_name).await?;
-    logger.log(
-        logging::LogLevel::Info,
-        &format!(
-            "[DEBUG] Index {}.{} validity: {}",
-            schema_name, index_name, index_is_valid
-        ),
-    );
-
-    // if the index is invalid, skip the reindexing.since reindexing an invalid index will cause duplicate entries in the index.
-    if !index_is_valid {
-        logger.log(
-            logging::LogLevel::Warning,
-            &format!(
-                "Index is invalid, skipping reindexing {}.{}",
-                schema_name, index_name
-            ),
-        );
-        // Save skipped record to logbook
-        let index_data = save::IndexData {
-            schema_name: schema_name.clone(),
-            index_name: index_name.clone(),
-            index_type: index_type.clone(),
-            reindex_status: "invalid_index".to_string(),
-            before_size: None,
-            after_size: None,
-            size_change: None,
-        };
-        save::save_index_info(&client, &index_data).await?;
-
-        return Ok(());
-    }
-
-    if reindexing_results.active_vacuum
-        || reindexing_results.active_pgreindexer
-        || (reindexing_results.inactive_replication_slots && !skip_inactive_replication_slots)
-        || (reindexing_results.sync_replication_connection && !skip_sync_replication_connection)
-    {
-        logger.log(
-            logging::LogLevel::Info,
-            &format!(
-                "[DEBUG] Skipping {}.{} due to reindexing conditions",
-                schema_name, index_name
-            ),
-        );
-        logger.log_index_skipped(
-            &schema_name,
-            &index_name,
-            "Active vacuum, pgreindexer or inactive replication slots detected",
-        );
-
-        // Save skipped record to logbook
-        let index_data = save::IndexData {
-            schema_name: schema_name.clone(),
-            index_name: index_name.clone(),
-            index_type: index_type.clone(),
-            reindex_status: "skipped".to_string(),
-            before_size: None,
-            after_size: None,
-            size_change: None,
-        };
-        save::save_index_info(&client, &index_data).await?;
-
-        return Ok(());
-    }
-
-    logger.log(
-        logging::LogLevel::Info,
-        &format!(
-            "[DEBUG] Executing reindex for {}.{}",
-            schema_name, index_name
-        ),
-    );
-
-    // Check for potential deadlock before executing reindex
-    check_and_handle_deadlock_risk(&client, &schema_name, &index_name, &shared_tracker, &logger)
-        .await?;
-
-    let start_time = std::time::Instant::now();
-    let result = client.execute(&reindex_sql, &[]).await;
-    let duration = start_time.elapsed();
-
-    match &result {
-        Ok(_) => {
-            logger.log(
-                logging::LogLevel::Info,
-                &format!(
-                    "[DEBUG] Reindex SQL executed successfully for {}.{} in {:?}",
-                    schema_name, index_name, duration
-                ),
-            );
-        }
-        Err(e) => {
-            logger.log(
-                logging::LogLevel::Error,
-                &format!(
-                    "[DEBUG] Reindex SQL failed for {}.{} after {:?}: {}",
-                    schema_name, index_name, duration, e
-                ),
-            );
-        }
-    }
-
-    result.context(format!(
-        "Failed to reindex index {}.{}",
-        schema_name, index_name
-    ))?;
-
-    // Get after size
-    let after_size = get_index_size(&client, &schema_name, &index_name).await?;
-    let size_change = after_size - before_size;
-
-    if verbose {
-        logger.log_index_size_info(before_size, after_size, size_change);
-    }
-
-    // Additional check: validate index integrity before saving
-    logger.log(
-        logging::LogLevel::Info,
-        &format!("[DEBUG] Waiting 5 seconds for index record to be written to table before validation for {}.{}", schema_name, index_name),
-    );
-    tokio::time::sleep(tokio::time::Duration::from_secs(5)).await;
-    logger.log(
-        logging::LogLevel::Info,
-        &format!(
-            "[DEBUG] Validating index integrity before saving for {}.{}",
-            schema_name, index_name
-        ),
-    );
-    let index_is_valid = validate_index_integrity(&client, &schema_name, &index_name).await?;
-    logger.log(
-        logging::LogLevel::Info,
-        &format!(
-            "[DEBUG] Final validation result for {}.{}: {}",
-            schema_name, index_name, index_is_valid
-        ),
-    );
-
-    if !index_is_valid {
-        logger.log(
-            logging::LogLevel::Info,
-            &format!(
-                "[DEBUG] Index validation failed for {}.{}",
-                schema_name, index_name
-            ),
-        );
-        logger.log_index_validation_failed(&schema_name, &index_name);
-
-        // Save failed validation record
-        let index_data = save::IndexData {
-            schema_name: schema_name.clone(),
-            index_name: index_name.clone(),
-            index_type: index_type.clone(),
-            reindex_status: "validation_failed".to_string(),
-            before_size: Some(before_size),
-            after_size: Some(after_size),
-            size_change: Some(size_change),
-        };
-        save::save_index_info(&client, &index_data).await?;
-
-        return Ok(());
-    }
-
-    // save the index info
-    logger.log(
-        logging::LogLevel::Info,
-        &format!(
-            "[DEBUG] Saving success record for {}.{}",
-            schema_name, index_name
-        ),
-    );
-    let index_data = save::IndexData {
-        schema_name: schema_name.clone(),
-        index_name: index_name.clone(),
-        index_type: index_type.clone(),
-        reindex_status: "success".to_string(),
-        before_size: Some(before_size),
-        after_size: Some(after_size),
-        size_change: Some(size_change),
-    };
-    save::save_index_info(&client, &index_data).await?;
-
-    logger.log(
-        logging::LogLevel::Info,
-        &format!(
-            "[DEBUG] Successfully completed reindex for {}.{}",
-            schema_name, index_name
-        ),
-    );
-    logger.log_index_success(&schema_name, &index_name);
-
-    // Remove table from shared tracker
-    remove_table_from_tracker(&client, &schema_name, &index_name, &shared_tracker, &logger).await?;
-
-    Ok(())
-}
-
-async fn set_session_parameters(
-    client: &tokio_postgres::Client,
-    maintenance_work_mem_gb: u64,
-    max_parallel_maintenance_workers: u64,
-    maintenance_io_concurrency: u64,
-) -> Result<()> {
-    // This function can be improved to set session parameters from the cli arguments.
-    // For now set the session parameters to 0.
-    client
-        .execute(queries::SET_STATEMENT_TIMEOUT, &[])
-        .await
-        .context("Set the statement timeout.")?;
-    client
-        .execute(queries::SET_IDLE_SESSION_TIMEOUT, &[])
-        .await
-        .context("Set the idle_session timeout.")?;
-    client
-        .execute(queries::SET_APPLICATION_NAME, &[])
-        .await
-        .context("Set the application name.")?;
-
-    // The following operation defines the maintenance work mem in GB provided by the user.
-    client
-        .execute(
-            format!(
-                "SET maintenance_work_mem TO '{}GB';",
-                maintenance_work_mem_gb
-            )
-            .as_str(),
-            &[],
-        )
-        .await
-        .context("Set the maintenance work mem.")?;
-
-    // Get current max_parallel_workers setting
-    let rows = client
-        .query(queries::GET_MAX_PARALLEL_WORKERS, &[])
-        .await
-        .context("Failed to get max_parallel_workers setting")?;
-
-    if let Some(row) = rows.first() {
-        let max_parallel_workers_str: String = row.get(0);
-        let max_parallel_workers: u64 = max_parallel_workers_str
-            .parse()
-            .context("Failed to parse max_parallel_workers value")?;
-
-        // Safety check: ensure max_parallel_maintenance_workers is less than max_parallel_workers/2
-        let safe_limit = max_parallel_workers / 2;
-        if max_parallel_maintenance_workers >= max_parallel_workers {
-            return Err(anyhow::anyhow!(
-                "max_parallel_maintenance_workers ({}) must be less than max_parallel_workers ({})",
-                max_parallel_maintenance_workers,
-                max_parallel_workers
-            ));
-        }
-
-        if max_parallel_maintenance_workers >= safe_limit {
-            return Err(anyhow::anyhow!(
-                "max_parallel_maintenance_workers ({}) must be less than max_parallel_workers/2 ({}) for safety",
-                max_parallel_maintenance_workers,
-                safe_limit
-            ));
-        }
-
-        // Set max_parallel_maintenance_workers
-        client
-            .execute(
-                format!(
-                    "SET max_parallel_maintenance_workers TO '{}';",
-                    max_parallel_maintenance_workers
-                )
-                .as_str(),
-                &[],
-            )
-            .await
-            .context("Set the max_parallel_maintenance_workers.")?;
-    } else {
-        return Err(anyhow::anyhow!(
-            "Failed to get max_parallel_workers setting"
-        ));
-    }
-
-    // Validate maintenance_io_concurrency (max 512)
-    if maintenance_io_concurrency > 512 {
-        return Err(anyhow::anyhow!(
-            "maintenance_io_concurrency ({}) must be 512 or less",
-            maintenance_io_concurrency
-        ));
-    }
-
-    // Set maintenance_io_concurrency
-    client
-        .execute(
-            format!(
-                "SET maintenance_io_concurrency TO '{}';",
-                maintenance_io_concurrency
-            )
-            .as_str(),
-            &[],
-        )
-        .await
-        .context("Set the maintenance_io_concurrency.")?;
-
-    Ok(())
-}
-
-// Create a new database connection with session parameters set
-async fn create_connection_with_session_parameters(
-    connection_string: &str,
-    maintenance_work_mem_gb: u64,
-    max_parallel_maintenance_workers: u64,
-    maintenance_io_concurrency: u64,
-) -> Result<tokio_postgres::Client> {
-    // Connect to PostgreSQL
-    let (client, connection) = tokio_postgres::connect(connection_string, NoTls)
-        .await
-        .context("Failed to connect to PostgreSQL")?;
-
-    // Spawn the connection to run in the background
-    tokio::spawn(async move {
-        if let Err(e) = connection.await {
-            eprintln!("Connection error: {}", e);
-        }
-    });
-
-    // Set session parameters for this connection
-    set_session_parameters(
-        &client,
-        maintenance_work_mem_gb,
-        max_parallel_maintenance_workers,
-        maintenance_io_concurrency,
-    )
-    .await?;
-
-    Ok(client)
 }
 
 fn get_password_from_pgpass(
@@ -1078,7 +354,10 @@ async fn main() -> Result<()> {
     );
 
     // Perform reindexing checks once and share results
-    logger.log(logging::LogLevel::Info, "Performing reindexing checks...");
+    logger.log(
+        logging::LogLevel::Info,
+        "Performing reindexing checks...",
+    );
     let reindexing_results = perform_reindexing_checks(&client).await?;
     let reindexing_results = Arc::new(reindexing_results);
 
@@ -1107,9 +386,7 @@ async fn main() -> Result<()> {
     let connection_string = Arc::new(connection_string);
 
     // Create shared table tracker
-    let shared_tracker = Arc::new(tokio::sync::Mutex::new(SharedTableTracker {
-        tables_being_reindexed: std::collections::HashMap::new(),
-    }));
+    let shared_tracker = Arc::new(tokio::sync::Mutex::new(SharedTableTracker::new()));
 
     // Create tasks for all indexes
     let mut tasks = Vec::new();
@@ -1131,7 +408,7 @@ async fn main() -> Result<()> {
         let maintenance_work_mem_gb = args.maintenance_work_mem_gb;
         let max_parallel_maintenance_workers = args.max_parallel_maintenance_workers;
         let maintenance_io_concurrency = args.maintenance_io_concurrency;
-
+        
         let task = tokio::spawn(async move {
             // Acquire permit from semaphore
             let _permit = semaphore.acquire().await.unwrap();
@@ -1192,4 +469,4 @@ async fn main() -> Result<()> {
     final_logger.log(logging::LogLevel::Success, "Reindex process completed");
 
     Ok(())
-}
+}
